requirements.txt
Flask>=2.2
requests>=2.31
Werkzeug>=2.2
python-dotenv>=1.0
gunicorn>=20.1.0
psutil>=5.9.0
